feat: updating the styling and code blocks
