--- code-examples/core_features/getting_started/10_getting_started.ts
@@ -1,5 +1,0 @@
-import * as Kilt from '@kiltprotocol/sdk-js'
-
-export async function main() {
-  await Kilt.disconnect()
-}
