+++ code-examples/core_features/getting_started/1_getting_started.ts
@@ -1,3 +1,3 @@
-export function main() {
+export async function main() {
   console.log('hello world')
 }
