+++ code-examples/core_features/getting_started/5_getting_started.ts
@@ -1,9 +1,6 @@
 import * as Kilt from '@kiltprotocol/sdk-js'
 
 export async function main(johnDoeDidId: string) {
- 
-  if (!johnDoeDidId) return console.log(`john_doe doesn't exist`)
-
   const johnDoeDid = await Kilt.Did.DidResolver.resolveDoc(johnDoeDidId)
 
   console.log(`John Doe's DID`, johnDoeDid)
@@ -11,6 +8,8 @@ export async function main(johnDoeDidId: string) {
   const endPoints = johnDoeDid?.details?.getEndpoints()
 
   if (!endPoints) return console.log('no endpoints')
+
+  console.log('Endpoints:', endPoints[0].urls[0])
 
   return endPoints
 }
