+++ code-examples/core_features/getting_started/7_getting_started.ts
@@ -1,8 +1,14 @@
 import * as Kilt from '@kiltprotocol/sdk-js'
 
 export async function main(request) {
-
   const attestation = await Kilt.Attestation.query(request.rootHash)
 
-  return attestation
+  const credential = Kilt.Credential.fromRequestAndAttestation(
+    request,
+    attestation
+  )
+
+  console.log('John Doe:', credential)
+
+  return credential
 }
